requirements.txt
canvasapi>=3.3.0
supabase>=2.3.0
pydantic>=2.5.0
httpx>=0.25.0
pytest>=7.4.0
python-dotenv>=1.0.0
pytz>=2023.3.post1
